新增了生成多分量信号训练集的功能

+++ getDataset.m
@@ -17,6 +17,7 @@
 %       Mode = 4:   为图像去噪网络的测试集生成时域信号
 %       Mode = 5:   为噪声估计网络的训练集生成时域信号
 %       Mode = 6:   为噪声估计网络的测试集生成时域信号
+%       Mode = 7:   为复合调制信号识别任务的训练集生成时域信号，并将每个类的样本存到单独的文件夹中
 %   SNRmin：信噪比的最小值
 %   SNRmax：信噪比的最大值
 %   Modulations: 25 elements represents NS, EQFM, LFM, VTFM, BPSK, QPSK, BFSK, SFM, COSTAS, TANFM, FRANK, P1, P2, P3, P4, T1, T2, T3, T4, LFM-SFM, EQFM-SFM, DDC-MASK, DDC-MPSK, DDC-MFSK, MQAM respectively. 1 means generate, 0 means do not generate
@@ -25,7 +26,6 @@
 %   [2] J. E. Fielding, "Polytime coding as a means of pulse compression," in IEEE Transactions on Aerospace and Electronic Systems, vol. 35, no. 2, pp. 716-721, April 1999, doi: 10.1109/7.766951.
 %   [3] https://www.radartutorial.eu/08.transmitters/Barker%20Code.en.html
 function [] = getDataset(varargin)
-addpath("tftb-0.2\mfiles\");
 disp("generating time domain signal ...");
 if nargin == 8
     Mode = varargin{1};
@@ -63,11 +63,7 @@ AllFiles = dir(pwd);
 fprintf('\n')
 disp("deleting existing folders ...");
 for i = 3:length(AllFiles) % delete all generated folders
-    if length(AllFiles(i).name) <= 2
-        disp('DANGER!')
-        return
-    end
-    if AllFiles(i).isdir == 1
+    if AllFiles(i).isdir == 1 && strcmp(AllFiles(i).name,".git") == 0 % 不可删除.git文件夹
         rmdir(AllFiles(i).name,'s')
     end
 end
@@ -571,4 +567,5 @@ if Modulations(25) == 1
 end
 delete(h);close all;
 assert(GlobalIndex==samples_num, "样本个数计算不正确！")
+getMultiComponentSignal(Mode,Modulations);
 end
